refactor: move two lodgers out of the files they were living in
Pure relocation. The diff on both source files is deletion only — no line of
logic was edited, and `xcodegen generate` picked the new files up.

**DashboardViewController.swift, 3260 → 1936.** It held two unrelated classes.
`DashboardOverviewView` is 1291 lines and refers to `DashboardViewController`
exactly zero times; the dependency runs one way (the controller names the view
53 times). Its two private helpers — `FlippedStackView`,
`NonFirstResponderScrollView` — are used only from inside it, as is the private
`FleetHoverRow` protocol, so all four move together into
`DashboardOverviewView.swift`.

`Array.subscript(safeIndex:)` sat at line 1954 of a view-controller file, under
a doc comment describing something else entirely, and nothing outside this file
used it. It gets `Array+SafeIndex.swift`, and the stranded doc comment is
reattached to the view it was always describing.

**MainWindowController.swift, 3126 → 3035.** Only `SeahelmWindow` moves — a
separate `NSWindow` subclass that happened to share the file.

The 1032 lines of extensions below it stay put, deliberately. Swift's `private`
is file-scoped, so moving them out means widening access on every class member
they touch: measured, that is 37 of the 96 private members, and only 20 of
those 1032 lines (`NewBranchDialogDelegate`, `WorktreeStatusDelegate`) are free
of them. Trading real encapsulation for a shorter file is not a refactor. The
seam worth taking there is the 539-line bridge/chat cluster inside the
mislabelled `// MARK: - Layout` block, which touches only 4 of 33 stored
properties — a collaborator to extract, not text to shuffle.
